+++ src/app/register/dietary-preference/dietary-preference.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import {NavigationService} from "../../navigation-service";
+import {NavigationService} from "../../services/navigation.service";
 
 @Component({
     selector: 'app-dietary-preference',
