+++ src/app/register/sign-up/sign-up.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core'
 import { FormBuilder } from '@angular/forms'
-import { SupabaseService} from "../../supabase.service";
+import { SupabaseService} from "../../services/supabase.service";
 
 
 @Component({
